refactor: create transaction authorization use case base
