+++ src/main/kotlin/domain/transaction/SimpleTransactionAuthorizationUseCase.kt
@@ -1,28 +1,16 @@
 package dev.emanuelmt.domain.transaction
 
+import dev.emanuelmt.domain.wallet.WalletEntity
 import dev.emanuelmt.domain.wallet.WalletRepository
-import dev.emanuelmt.domain.wallet.subBalance
 
 class SimpleTransactionAuthorizationUseCase(
-    private val transactionRepository: TransactionRepository,
+    transactionRepository: TransactionRepository,
     private val walletRepository: WalletRepository
-) {
-    suspend fun execute(input: TransactionAuthorizationInput): TransactionAuthorizationOutput {
+): TransactionAuthorizationUseCaseBase(transactionRepository, walletRepository) {
+
+    override suspend fun determineWallet(input: TransactionAuthorizationInput): WalletEntity? {
         val balanceType = merchantTypeToBalanceType(input.merchantType)
-
-        try {
-            walletRepository.find(input.accountId, balanceType)?.let { wallet ->
-                if (wallet.balance < input.amount) return TransactionAuthorizationOutput(TransactionStatusCode.InsufficientBalance)
-                wallet.subBalance(input.amount).also {
-                    val transaction = newTransaction(it.id, input.merchantType, input.merchantName, input.amount)
-                    transactionRepository.save(transaction) { walletRepository.update(it) }
-                    return TransactionAuthorizationOutput(TransactionStatusCode.ApprovedTransaction)
-                }
-            }
-        } catch (exception: Exception) {
-            println(exception)
-        }
-
-        return TransactionAuthorizationOutput(TransactionStatusCode.RejectedTransaction)
+        return walletRepository.find(input.accountId, balanceType)
     }
+
 }
